+++ project/src/components/tab-detail/tab-detail.tsx
@@ -1,0 +1,47 @@
+import {Film} from '../../types/film';
+
+type TabDetailProps = {
+  film: Film;
+};
+
+export default function TabDetail ({film}: TabDetailProps): JSX.Element {
+  const {
+    director,
+    starring,
+    runTime,
+    genre,
+    released
+  } = film;
+
+  return (
+    <div className="film-card__text film-card__row">
+      <div className="film-card__text-col">
+        <p className="film-card__details-item">
+          <strong className="film-card__details-name">Director</strong>
+          <span className="film-card__details-value">{director}</span>
+        </p>
+        <p className="film-card__details-item">
+          <strong className="film-card__details-name">Starring</strong>
+          <span className="film-card__details-value">
+            {starring.join(', ')}
+          </span>
+        </p>
+      </div>
+
+      <div className="film-card__text-col">
+        <p className="film-card__details-item">
+          <strong className="film-card__details-name">Run Time</strong>
+          <span className="film-card__details-value">{runTime}</span>
+        </p>
+        <p className="film-card__details-item">
+          <strong className="film-card__details-name">Genre</strong>
+          <span className="film-card__details-value">{genre}</span>
+        </p>
+        <p className="film-card__details-item">
+          <strong className="film-card__details-name">Released</strong>
+          <span className="film-card__details-value">{released}</span>
+        </p>
+      </div>
+    </div>
+  );
+}
